+++ src/Data.php
@@ -2,83 +2,29 @@
 
 namespace h4kuna\Ares;
 
+use h4kuna\DataType\Immutable;
+
 /**
  * @author Milan Matějček
+ * @property-read bool $active
+ * @property-read string $city
+ * @property-read string $city_district
+ * @property-read string $city_post
+ * @property-read string $company
+ * @property-read string $court
+ * @property-read string $court_all
+ * @property-read \DateTime $created
+ * @property-read string $file_number
+ * @property-read string $in
+ * @property-read bool $is_person
+ * @property-read string $house_number
+ * @property-read string $street
+ * @property-read string $tin
+ * @property-read string $vat_payer
+ * @property-read string $zip
  */
-class Data implements IData, \ArrayAccess, \Iterator, \Countable
+class Data extends Immutable\Messenger
 {
-
-	private $data = [];
-
-	/**
-	 * @param bool $isActive
-	 */
-	public function setActive($isActive)
-	{
-		$this->data['active'] = $isActive;
-	}
-
-	public function setCity($city)
-	{
-		$this->data['city'] = $city;
-	}
-
-	public function setCompany($company)
-	{
-		$this->data['company'] = $company;
-	}
-
-	public function setCourt($court)
-	{
-		$this->data['court'] = $court;
-	}
-
-	public function setCreated(\DateTime $date)
-	{
-		$this->data['created'] = $date;
-	}
-
-	public function setFileNumber($fileNumber)
-	{
-		$this->data['file_number'] = $fileNumber;
-	}
-
-	public function setPerson($isPerson)
-	{
-		$this->data['person'] = $isPerson;
-	}
-
-	public function setStreet($street)
-	{
-		$this->data['street'] = $street;
-	}
-
-	public function setIN($in)
-	{
-		$this->data['in'] = $in;
-	}
-
-	public function setTIN($tin)
-	{
-		$this->data['tin'] = $tin;
-	}
-
-	public function setVatpayer($vatPay)
-	{
-		$this->data['vat_payer'] = $vatPay;
-	}
-
-	public function setZip($zip)
-	{
-		$this->data['zip'] = $zip;
-	}
-
-	private function setFileNumberAndCourt()
-	{
-		if (!isset($this->data['court_all']) && array_key_exists('file_number', $this->data) && array_key_exists('court', $this->data)) {
-			$this->data['court_all'] = $this->data['file_number'] . ', ' . $this->data['court'];
-		}
-	}
 
 	/**
 	 * Copy data
@@ -87,9 +33,8 @@ class Data implements IData, \ArrayAccess, \Iterator, \Countable
 	 */
 	public function toArray(array $map = [])
 	{
-		$this->setFileNumberAndCourt();
 		if ($map === []) {
-			return $this->data;
+			return $this->getData();
 		}
 		$out = [];
 		foreach ($map as $k => $v) {
@@ -97,126 +42,23 @@ class Data implements IData, \ArrayAccess, \Iterator, \Countable
 				if (!$v) {
 					$v = $k;
 				}
-				$out[$v] = $this->data[$k];
+				$out[$v] = $this[$k];
 			}
 		}
 		return $out;
 	}
 
+	public function jsonSerialize()
+	{
+		$data = $this->getData();
+		if ($this->created instanceof \DateTime) {
+			$data['created'] = $this->created->format(\DateTime::ISO8601);
+		}
+		return $data;
+	}
+
 	public function __toString()
 	{
-		$data = $this->toArray();
-		if ($data['created'] instanceof \DateTime) {
-			$data['created'] = $data['created']->format(\DateTime::ISO8601);
-		}
-		return json_encode($data);
+		return json_encode($this->jsonSerialize());
 	}
-
-	/**
-	 * ARRAY-ACCESS INTERFACE **************************************************
-	 * *************************************************************************
-	 */
-
-	/**
-	 *
-	 * @param string $offset
-	 * @return bool
-	 */
-	public function offsetExists($offset)
-	{
-		return array_key_exists($offset, $this->data);
-	}
-
-	/**
-	 * Return value from array
-	 * @param string $offset
-	 * @return string
-	 * @throws DataOffsetDoesNotExists
-	 */
-	public function offsetGet($offset)
-	{
-		if ($offset === 'court_all') {
-			$this->setFileNumberAndCourt();
-		}
-		if ($this->offsetExists($offset)) {
-			return $this->data[$offset];
-		}
-		throw new DataOffsetDoesNotExists($offset);
-	}
-
-	/**
-	 * @param string $offset
-	 * @param string $value
-	 * @return string
-	 */
-	public function offsetSet($offset, $value)
-	{
-		return $this->data[$offset] = $value;
-	}
-
-	/**
-	 * Remove value from array
-	 * @param string $offset
-	 * @return void
-	 */
-	public function offsetUnset($offset)
-	{
-		unset($this->data[$offset]);
-	}
-
-	/**
-	 * ITERATOR INTERFACE ******************************************************
-	 * *************************************************************************
-	 */
-
-	/**
-	 * Actual value
-	 * @return void
-	 */
-	public function current()
-	{
-		return current($this->data);
-	}
-
-	/**
-	 * Actual key of value
-	 * @return string
-	 */
-	public function key()
-	{
-		return key($this->data);
-	}
-
-	/**
-	 * Next value
-	 * @return string
-	 */
-	public function next()
-	{
-		return next($this->data);
-	}
-
-	/** @retrun void */
-	public function rewind()
-	{
-		reset($this->data);
-	}
-
-	/** @return bool */
-	public function valid()
-	{
-		return array_key_exists($this->key(), $this->data);
-	}
-
-	/**
-	 * COUNTABLE INTERFACE *****************************************************
-	 * *************************************************************************
-	 */
-
-	/** @return int */
-	public function count()
-	{
-		return count($this->data);
-	}
-
 }